traza11 .intValue()

+++ servicio-modelosbalizas/src/main/java/com/galileo/cu/servicio_modelosbalizas/controladores/ModelosBalizasEventHandler.java
@@ -78,7 +78,7 @@ public class ModelosBalizasEventHandler {
         try {
             String descripcionTraza = "Fue creado un nuevo modelo: " + modelo.getDescripcion();
             log.info("*****HandleAfterCreate MODELOS BALIZAS*****");
-            ActualizarTraza(val, modelo.getId(), 1, 1, descripcionTraza,
+            ActualizarTraza(val, modelo.getId().intValue(), 1, 1, descripcionTraza,
                     "Fallo al Actualizar el Modelo en la Trazabilidad");
 
         } catch (Exception e) {
@@ -126,7 +126,7 @@ public class ModelosBalizasEventHandler {
         try {
             String descripcionTraza = "Fue Actualizado un nuevo modelo: " + modelo.getDescripcion();
             log.info("*****handleAfterUpdate MODELOS BALIZAS*****");
-            ActualizarTraza(val, modelo.getId(), 1, 3, descripcionTraza,
+            ActualizarTraza(val, modelo.getId().intValue(), 1, 3, descripcionTraza,
                     "Fallo al Actualizar el Modelo en la Trazabilidad");
 
         } catch (Exception e) {
